Changed indicator to nullable

+++ blueprint-automationframework/Blueprint-Tests/Tests/ArtifactStoreTests/AttachmentTests.cs
@@ -86,7 +86,7 @@ namespace ArtifactStoreTests
             Assert.AreEqual(attachment.AttachedFiles[0].UploadedDate, artifactDetails.LastEditedOn,
                 "UploadedDate for published artifact's attachment should be equal to LastEditedOn date of artifact");
 
-            ArtifactStoreHelper.VerifyIndicatorFlags(Helper, _adminUser, artifact.Id, (int)ItemIndicatorFlags.HasAttachmentsOrDocumentRefs);
+            ArtifactStoreHelper.VerifyIndicatorFlags(Helper, _adminUser, artifact.Id, ItemIndicatorFlags.HasAttachmentsOrDocumentRefs);
         }
 
         [TestCase]
